Added custom exceptions related to BookService and modified version of springboot-starter

+++ library-backend/src/main/java/com/botond/librarybackend/service/BookService.java
@@ -1,6 +1,8 @@
 package com.botond.librarybackend.service;
 
 import com.botond.librarybackend.entity.Book;
+import com.botond.librarybackend.error.BookAlreadyExistsException;
+import com.botond.librarybackend.error.BookNotFoundException;
 import com.botond.librarybackend.repository.BookRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,11 +22,24 @@ public class BookService {
     }
 
     public void addBook(Book newBook) {
+        List<Book> books = getAllBook();
+        for (Book book : books) {
+            if(newBook.getISBN().equals(book.getISBN())) {
+                throw new BookAlreadyExistsException(newBook.getISBN());
+            }
+        }
         bookRepository.save(newBook);
     }
 
     public void deleteBook(Long Id) {
-        bookRepository.deleteById(Id);
+        List<Book> books = getAllBook();
+        for (Book book : books) {
+            if(book.getId().equals(Id)) {
+                bookRepository.deleteById(Id);
+                return;
+            }
+        }
+        throw new BookNotFoundException(Id);
     }
 
     public void updateQuantity(Long Id, Long newQuantity) {
@@ -32,7 +47,8 @@ public class BookService {
                 .map(x -> {
                     x.setQuantity(newQuantity);
                     return bookRepository.save(x);
-                    });
+                    })
+        .orElseThrow(() -> new BookNotFoundException(Id));
     }
 
     public void updateDescription(Long Id, String newDesc) {
@@ -40,6 +56,7 @@ public class BookService {
                 .map(x -> {
                     x.setDescription(newDesc);
                     return bookRepository.save(x);
-                });
+                })
+        .orElseThrow(() -> new BookNotFoundException(Id));
     }
 }
